+++ ChangeLog.md
@@ -6,6 +6,8 @@
 * Introducing module Data.Verbosity.Class which contains definition of
   HasVerbosity type class. (new)
 * Introducing function `fromInt :: Int -> Maybe Verbosity`. (new)
+* Introducing function
+  `parse :: (Eq string, IsString string) => string -> Maybe Verbosity`. (new)
 * Depends on [transformers][] package in case [base][] <4.8. (new)